+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 - Fail gracefully if no ssoCard is present for asteroid when querying datacloud catalogue
 - Added weighted average output to some datacloud catalogue queries from the command line
 - Entries where both the property and its error are zero in datacloud catalogues are no longer preferred
+- Print warning when no ssoCard could be retrieved for an asteroid.
+- Added 'no_cache' keyword to 'rocks.ssodnet.get_ssocard'. If True, it forces the remote query of the ssoCard
+- Improved 'rocks status' command to echo the number of cached ssoCards and offer to update the out-of-date ones
+- Fixed typo in ssodnet name of astorb catalogue
 
 # 1.2.2 - 2021-08-16
 - Fixed unit lookup file